The FrameControl dialog has been coming up in LessTif pushed way down to the bottom of the dialog. This would seem to be a problem with LessTif, but a 2 line fix of dx seems easier right now. If you don't like it you try to figure out how XmForm is supposed to track children's preferred size!!!!
The FrameControl is creating the SlideBar with XmNheight equal to the height
of the FrameControl and a topPosition of 47. This is nuts. It turns out that
the FrameControl wants to be 180 high, so it is creating the SlideBar 180
high. This is what LessTif is doing, exactly what it was told to. Where
exactly Motif is deciding that the programmer is wacked I'm not too sure, but
simply letting the toolkit do it's job and calculate the height it needs
fixes the problem once again.