A generous sprinkling of ${top_srcdir} and ${srcdir} all over the place to get this puppy building out of the source tree. You can now (or at least I can) check this beast out of cvs, CVSMake it, burn the result on a cdrom then carry this around and build it where ever you want. Or build multiple arch's off of the same source tree.
Something to keep in mind.....any files referenced in the source tree should
be prefixed with ${top_srcdir}, ${srcdir}, or one of the other autoconf
defined symbols. If you need to refer to a file that is made from a .in file,
it will be relative to ${top_builddir}.

Some things may still barf if srcdir != builddir, for example, the java stuff
which I do not configure. If you need it, it should be pretty simple to fix
up by looking at some of these files. Unless it's like src/exec/libdx/*,
that one was a bit of a b.....

While I have your attention, have I been "black balled" from the opendx-dev
list?? I'm getting the traffic on the list, but I have replied to a couple of
messages in the last couple of days, and haven't seen any of my replies on
the list.