In the past, SMP OpenDX has always forked n times for n-way parallelism, with the original process really doing nothing but waiting for the kids to finish.  Further, the master process - eg. DXProcessorId() == 0, is the last one forked, not the first.
In SMP OpenDX MPI, it is necessary that all MPI communication comes from
one process, and further, that it be the initial process created by the
parallel operating environment.  For this reason I've fixed the code that
is in there to avoid this wait process (see the DXD_EXEC_WAIT_PROCESS
#define) and added another - DXD_PROC_0_FORKED_FIRST.  By defining
DXD_EXEC_WAIT_PROCESS = 0 and DXD_PROC_0_FORKED_FIRST = 1, you get the
initial process acting as the master.

Didn't change the default behavior.