For OpenGL (hardware) rendering mode, adds the following field attributes that affect culling and lighting:
     attribute "cull face" string ["off"|"front"|"back"|"front and back"]
     attribute "light model" string ["one side"|"two side"]

and the following field attributes that affect texturing (when the
"texture" attribute and "uv" component are also correctly provided of course):

     attribute "texture wrap s" string ["clamp"|"repeat"]
     attribute "texture wrap t" string ["clamp"|"repeat"]
     attribute "texture min filter" string \
         ["nearest"|"linear"|"nearest_mipmap_nearest"|"nearest_mipmap_linear"|
          "linear_mipmap_nearest"|"linear_mipmap_linear"]
     attribute "texture mag filter" string ["nearest"|"linear"]
     attribute "texture function" string ["decal"|"replace"|"modulate"|"blend"]

     Note that "light model = two side" is particularly useful when
lighting polygonal models with no normals where the auto-computed normal
may bear no resemblence to the "outside" of a rendered object (without it,
only software rendering in DX is worthwhile).  The rest are useful when 3D
polygonal models that require specific culling and texturing state for
correct display are integrated into DX visualizations.

     Note that the default OpenGL state when these attributes are not
specified is the same as it was before.

     Also, now that texture maps are registered via gluBuild2DMipmaps in
the OpenGL back-end, the restrictions on supplied texture maps being powers
of two and larger than >= 64x64 is no longer the case in general, so the
checks have been removed.  Back-end specific checks can be added later if
other back-ends are still used.