Better ABI detection and handling for SGI.
Adds one new configure-generated (dxconfig.h) variable: SGI_ABI, which on
IRIX is defined to be one of the following: -32, -n32, -64.  Detection is
from the compiler-defined C preprocessor symbol _MIPS_SIM, so this covers
any of the half dozen ways you can set the target ABI for the IRIX C/C++
compilers.  Coincidentally, this setting is the same as the values for the
IRIX env var $SGI_ABI and the corresponding C/C++ compiler options used to
set the target ABI.

Also, changes the setting of the DXABI variable (used by Java) to be
completely based on this value.  There was no real testing before -- just a
default applicable to n32.